Version control
Jupyter Notebook files contain rich media, metadata, and formatted text which makes them not work so well with normal version control methods. 
The aim is to avoid this by resetting notebook files so that no outputs are included, leaving just the source code to download and upload to github.
This will provide a way of differentiating between versions/ analysing individual update changes.